+++ example/index.js
@@ -27,6 +27,7 @@ import { LDrawUtils } from 'three/examples/jsm/utils/LDrawUtils.js';
 import { GUI } from 'three/examples/jsm/libs/lil-gui.module.min.js';
 import { PathTracingViewer } from '../src/viewers/PathTracingViewer.js';
 import Stats from 'three/examples/jsm/libs/stats.module.js';
+import { generateRadialFloorTexture } from './utils/generateRadialFloorTexture.js';
 
 const creditEl = document.getElementById( 'credits' );
 const loadingEl = document.getElementById( 'loading' );
@@ -441,45 +442,6 @@ function buildGui() {
 
 	} );
 	pathTracingFolder.open();
-
-}
-
-function generateRadialFloorTexture( dim ) {
-
-	const data = new Uint8Array( dim * dim * 4 );
-
-	for ( let x = 0; x < dim; x ++ ) {
-
-		for ( let y = 0; y < dim; y ++ ) {
-
-			const xNorm = x / ( dim - 1 );
-			const yNorm = y / ( dim - 1 );
-
-			const xCent = 2.0 * ( xNorm - 0.5 );
-			const yCent = 2.0 * ( yNorm - 0.5 );
-			let a = Math.max( Math.min( 1.0 - Math.sqrt( xCent ** 2 + yCent ** 2 ), 1.0 ), 0.0 );
-			a = a ** 2;
-			a = Math.min( a, 1.0 );
-
-			const i = y * dim + x;
-			data[ i * 4 + 0 ] = 255;
-			data[ i * 4 + 1 ] = 255;
-			data[ i * 4 + 2 ] = 255;
-			data[ i * 4 + 3 ] = a * 255;
-
-		}
-
-	}
-
-	const tex = new DataTexture( data, dim, dim );
-	tex.format = RGBAFormat;
-	tex.type = UnsignedByteType;
-	tex.minFilter = LinearFilter;
-	tex.magFilter = LinearFilter;
-	tex.wrapS = RepeatWrapping;
-	tex.wrapT = RepeatWrapping;
-	tex.needsUpdate = true;
-	return tex;
 
 }
