+++ example/skinnedMesh.js
@@ -2,10 +2,11 @@ import * as THREE from 'three';
 import { FullScreenQuad } from 'three/examples/jsm/postprocessing/Pass.js';
 import { GLTFLoader } from 'three/examples/jsm/loaders/GLTFLoader.js';
 import { OrbitControls } from 'three/examples/jsm/controls/OrbitControls.js';
-import { PathTracingPersistentSceneGenerator, PathTracingRenderer, PhysicalPathTracingMaterial } from '../src/index.js';
+import { DynamicPathTracingSceneGenerator, PathTracingRenderer, PhysicalPathTracingMaterial } from '../src/index.js';
 import { RGBELoader } from 'three/examples/jsm/loaders/RGBELoader.js';
 import { MeshoptDecoder } from 'three/examples/jsm/libs/meshopt_decoder.module.js';
 import { GUI } from 'three/examples/jsm/libs/lil-gui.module.min.js';
+import { generateRadialFloorTexture } from './utils/generateRadialFloorTexture.js';
 
 let renderer, controls, sceneInfo, ptRenderer, camera, fsQuad, scene, clock, model;
 let samplesEl;
@@ -162,6 +163,7 @@ function loadModel( url ) {
 		.loadAsync( url )
 		.then( gltf => {
 
+			// make the model white since the texture seems to dark for the env map
 			gltf.scene.traverse( c => {
 
 				if ( c.material ) {
@@ -173,9 +175,6 @@ function loadModel( url ) {
 
 			} );
 
-			const group = new THREE.Group();
-			group.add( gltf.scene );
-
 			// animations
 			const animations = gltf.animations;
 			const mixer = new THREE.AnimationMixer( gltf.scene );
@@ -183,6 +182,10 @@ function loadModel( url ) {
 			const action = mixer.clipAction( animations[ 0 ] );
 			action.play();
 			action.paused = params.pause;
+
+			// add floor
+			const group = new THREE.Group();
+			group.add( gltf.scene );
 
 			const floorTex = generateRadialFloorTexture( 2048 );
 			const floorPlane = new THREE.Mesh(
@@ -200,8 +203,9 @@ function loadModel( url ) {
 			floorPlane.position.y = 0.075;
 			group.add( floorPlane );
 
-			const sceneGenerator = new PathTracingPersistentSceneGenerator( group );
-			sceneGenerator.synchronous = true;
+			// create the scene generator for updating skinned meshes quickly
+			const sceneGenerator = new DynamicPathTracingSceneGenerator( group );
+
 			return {
 				scene: group,
 				sceneGenerator,
@@ -215,44 +219,6 @@ function loadModel( url ) {
 
 }
 
-function generateRadialFloorTexture( dim ) {
-
-	const data = new Uint8Array( dim * dim * 4 );
-
-	for ( let x = 0; x < dim; x ++ ) {
-
-		for ( let y = 0; y < dim; y ++ ) {
-
-			const xNorm = x / ( dim - 1 );
-			const yNorm = y / ( dim - 1 );
-
-			const xCent = 2.0 * ( xNorm - 0.5 );
-			const yCent = 2.0 * ( yNorm - 0.5 );
-			let a = Math.max( Math.min( 1.0 - Math.sqrt( xCent ** 2 + yCent ** 2 ), 1.0 ), 0.0 );
-			a = a ** 2;
-			a = Math.min( a, 1.0 );
-
-			const i = y * dim + x;
-			data[ i * 4 + 0 ] = 255;
-			data[ i * 4 + 1 ] = 255;
-			data[ i * 4 + 2 ] = 255;
-			data[ i * 4 + 3 ] = a * 255;
-
-		}
-
-	}
-
-	const tex = new THREE.DataTexture( data, dim, dim );
-	tex.format = THREE.RGBAFormat;
-	tex.type = THREE.UnsignedByteType;
-	tex.minFilter = THREE.LinearFilter;
-	tex.magFilter = THREE.LinearFilter;
-	tex.wrapS = THREE.RepeatWrapping;
-	tex.wrapT = THREE.RepeatWrapping;
-	tex.needsUpdate = true;
-	return tex;
-
-}
 
 function onResize() {
 
@@ -351,7 +317,3 @@ function animate() {
 	}
 
 }
-
-
-
-
