+++ example/utils/generateRadialFloorTexture.js
@@ -1,0 +1,39 @@
+
+export function generateRadialFloorTexture( dim ) {
+
+	const data = new Uint8Array( dim * dim * 4 );
+
+	for ( let x = 0; x < dim; x ++ ) {
+
+		for ( let y = 0; y < dim; y ++ ) {
+
+			const xNorm = x / ( dim - 1 );
+			const yNorm = y / ( dim - 1 );
+
+			const xCent = 2.0 * ( xNorm - 0.5 );
+			const yCent = 2.0 * ( yNorm - 0.5 );
+			let a = Math.max( Math.min( 1.0 - Math.sqrt( xCent ** 2 + yCent ** 2 ), 1.0 ), 0.0 );
+			a = a ** 2;
+			a = Math.min( a, 1.0 );
+
+			const i = y * dim + x;
+			data[ i * 4 + 0 ] = 255;
+			data[ i * 4 + 1 ] = 255;
+			data[ i * 4 + 2 ] = 255;
+			data[ i * 4 + 3 ] = a * 255;
+
+		}
+
+	}
+
+	const tex = new THREE.DataTexture( data, dim, dim );
+	tex.format = THREE.RGBAFormat;
+	tex.type = THREE.UnsignedByteType;
+	tex.minFilter = THREE.LinearFilter;
+	tex.magFilter = THREE.LinearFilter;
+	tex.wrapS = THREE.RepeatWrapping;
+	tex.wrapT = THREE.RepeatWrapping;
+	tex.needsUpdate = true;
+	return tex;
+
+}
